requirements.txt
requests>=2.25.1
pyjwt>=2.0.0 